backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery[redis]==5.3.4
redis==4.6.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.1
numpy==1.24.3
pandas==2.1.3
pydantic==2.5.0
pydantic-settings==2.1.0
pgvector==0.2.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.6.1
groq==0.4.1
aiofiles==23.2.1
python-dotenv==1.0.0
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
SpeechRecognition==3.10.1
pyaudio==0.2.14
pydub==0.25.1